requirements.txt
requests_html
beautifulsoup4
python-dotenv
lxml_html_clean
streamlit
pandas
pydeck
geopy